rename: merge 후 폴더 경로 수정 [TRL-39, TRL-40]
